hidalgo_mx_chatbot_twin/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pdfminer.six
torch
transformers
accelerate
nest_asyncio
